Added better pricing strings

+++ gthockey-iOS/gthockey-iOS/JSONParser.swift
@@ -183,7 +183,7 @@ class JSONParser {
     private func makeApparelObject(value: JSON) -> Apparel {
         let apparel = Apparel(id: value["id"].int!,
                               name: value["name"].string!,
-                              price: value["price"].float!,
+                              price: value["price"].double!,
                               description: value["description"].string!,
                               imageURL: URL(string: value["image"].string ?? "https://test.gthockey.com/media/players/caleb.jpg")!)
         return apparel
